+++ apple-codesign/src/macos.rs
@@ -7,19 +7,29 @@
 use {
     crate::{
         certificate::{AppleCertificate, OID_USER_ID},
+        cryptography::PrivateKey,
         error::AppleCodesignError,
+        remote_signing::{session_negotiation::PublicKeyPeerDecrypt, RemoteSignError},
     },
     bcder::Oid,
+    bytes::Bytes,
+    log::{error, warn},
     security_framework::{
         certificate::SecCertificate,
         item::{ItemClass, ItemSearchOptions, Reference, SearchResult},
+        key::{Algorithm as KeychainAlgorithm, SecKey},
         os::macos::{
+            encrypt_transform::{Builder as EncryptBuilder, Padding},
             item::ItemSearchOptionsExt,
             keychain::{SecKeychain, SecPreferencesDomain},
         },
     },
+    signature::Signer,
     std::ops::Deref,
-    x509_certificate::CapturedX509Certificate,
+    x509_certificate::{
+        CapturedX509Certificate, KeyAlgorithm, KeyInfoSigner, Sign, Signature, SignatureAlgorithm,
+        X509CertificateError,
+    },
 };
 
 const SYSTEM_ROOTS_KEYCHAIN: &str = "/System/Library/Keychains/SystemRootCertificates.keychain";
@@ -62,8 +72,10 @@ impl TryFrom<&str> for KeychainDomain {
 }
 
 /// A certificate in a keychain.
+#[derive(Clone)]
 pub struct KeychainCertificate {
     sec_cert: SecCertificate,
+    sec_key: SecKey,
     captured: CapturedX509Certificate,
 }
 
@@ -75,12 +87,121 @@ impl Deref for KeychainCertificate {
     }
 }
 
+impl Signer<Signature> for KeychainCertificate {
+    fn try_sign(&self, message: &[u8]) -> Result<Signature, signature::Error> {
+        let algorithm = self
+            .signature_algorithm()
+            .map_err(signature::Error::from_source)?;
+
+        let algorithm = match algorithm {
+            SignatureAlgorithm::RsaSha1 => KeychainAlgorithm::RSASignatureMessagePKCS1v15SHA1,
+            SignatureAlgorithm::RsaSha256 => KeychainAlgorithm::RSASignatureMessagePKCS1v15SHA256,
+            SignatureAlgorithm::RsaSha384 => KeychainAlgorithm::RSASignatureMessagePKCS1v15SHA384,
+            SignatureAlgorithm::RsaSha512 => KeychainAlgorithm::RSASignatureMessagePKCS1v15SHA512,
+            SignatureAlgorithm::EcdsaSha256 => KeychainAlgorithm::ECDSASignatureMessageX962SHA256,
+            SignatureAlgorithm::EcdsaSha384 => KeychainAlgorithm::ECDSASignatureMessageX962SHA384,
+            SignatureAlgorithm::Ed25519 => KeychainAlgorithm::ECDSASignatureMessageX962SHA512,
+        };
+
+        warn!(
+            "attempting to create signature using keychain item: {}",
+            self.sec_cert.subject_summary()
+        );
+
+        let signature = self
+            .sec_key
+            .create_signature(algorithm, message)
+            .map_err(|e| {
+                signature::Error::from_source(format!(
+                    "when attempting to create signature from keychain item: {}",
+                    e
+                ))
+            })?;
+
+        Ok(Signature::from(signature))
+    }
+}
+
+impl Sign for KeychainCertificate {
+    fn sign(&self, message: &[u8]) -> Result<(Vec<u8>, SignatureAlgorithm), X509CertificateError> {
+        let algorithm = self.signature_algorithm()?;
+
+        Ok((self.try_sign(message)?.into(), algorithm))
+    }
+
+    fn key_algorithm(&self) -> Option<KeyAlgorithm> {
+        self.captured.key_algorithm()
+    }
+
+    fn public_key_data(&self) -> Bytes {
+        self.captured.public_key_data()
+    }
+
+    fn signature_algorithm(&self) -> Result<SignatureAlgorithm, X509CertificateError> {
+        Ok(self.captured.signature_algorithm().ok_or(
+            X509CertificateError::UnknownSignatureAlgorithm(format!(
+                "{:?}",
+                self.captured.signature_algorithm_oid()
+            )),
+        )?)
+    }
+
+    fn private_key_data(&self) -> Option<Vec<u8>> {
+        None
+    }
+
+    fn rsa_primes(&self) -> Result<Option<(Vec<u8>, Vec<u8>)>, X509CertificateError> {
+        Ok(None)
+    }
+}
+
+impl KeyInfoSigner for KeychainCertificate {}
+
+impl PublicKeyPeerDecrypt for KeychainCertificate {
+    fn decrypt(&self, _ciphertext: &[u8]) -> Result<Vec<u8>, RemoteSignError> {
+        // It doesn't look like the Rust bindings expose the APIs we need to
+        // implement decryption. Sadness. Will probably need to contribute
+        // those upstream...
+        error!("missing feature along with workarounds tracked in https://github.com/indygreg/PyOxidizer/issues/554");
+        Err(RemoteSignError::Crypto(
+            "decryption not yet implemented for keychain stored keys".into(),
+        ))
+    }
+}
+
+impl PrivateKey for KeychainCertificate {
+    fn as_key_info_signer(&self) -> &dyn KeyInfoSigner {
+        self
+    }
+
+    fn to_public_key_peer_decrypt(
+        &self,
+    ) -> Result<Box<dyn PublicKeyPeerDecrypt>, AppleCodesignError> {
+        Ok(Box::new(self.clone()))
+    }
+
+    fn finish(&self) -> Result<(), AppleCodesignError> {
+        Ok(())
+    }
+}
+
+impl KeychainCertificate {
+    /// Obtain a new [CapturedX509Certificate] for this item.
+    pub fn as_captured_x509_certificate(&self) -> CapturedX509Certificate {
+        self.captured.clone()
+    }
+}
+
 fn find_certificates(
     keychains: &[SecKeychain],
 ) -> Result<Vec<KeychainCertificate>, AppleCodesignError> {
     let mut search = ItemSearchOptions::default();
     search.keychains(keychains);
-    search.class(ItemClass::certificate());
+    // We fetch identities here because that gives us access to both the public
+    // cert and private key. The keychain doesn't need to be unlocked to get a
+    // handle on the private key: only when an operation on the private key is
+    // requested.
+    search.class(ItemClass::identity());
     search.limit(i32::MAX as i64);
 
     let mut certs = vec![];
@@ -88,10 +209,14 @@ fn find_certificates(
     for item in search.search()? {
         match item {
             SearchResult::Ref(reference) => match reference {
-                Reference::Certificate(cert) => {
+                Reference::Identity(identity) => {
+                    let cert = identity.certificate()?;
+                    let private_key = identity.private_key()?;
+
                     if let Ok(captured) = CapturedX509Certificate::from_der(cert.to_der()) {
                         certs.push(KeychainCertificate {
                             sec_cert: cert,
+                            sec_key: private_key,
                             captured,
                         });
                     }